# api/openapi.yaml
openapi: 3.0.3
info:
  title: Saga Orchestration Engine REST API
  description: >-
    HTTP/REST API for the go-saga-orchestration engine (the `api` binary, served on
    port 8080 by default). Covers health probes, saga lifecycle (start, signal,
    user-task submit, live stream, inspect), the action registry, rule
    evaluation, trigger CRUD, and workflow stats.


    All endpoints emit a structured JSON error envelope `{ "error": code,
    "message": msg }` on error. Error `code` is a stable, machine-readable
    string (e.g. `bad_request`, `not_found`, `internal`). On 5xx the real
    error is logged server-side; clients receive only the generic message.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local api binary (env WORKFLOW_API_PORT, default 8080)
paths:
...
  /api/v1/sagas/{run_id}/user_task/{task_id}/submit:
    post:
      summary: Submit a user task result
      description: >-
        Persists the task submission, appends a `user_task.{task_id}.submitted`
        signal carrying the result, and advances the saga if it was awaiting it.
      operationId: submitUserTask
      tags: [sagas]
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: task_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserTaskSubmitRequest"
            example:
              submitted_by: alice@example.com
              result: { decision: approve, comment: "LGTM" }
      responses:
        "202":
          description: Submission accepted. No body.
        "400":
          $ref: "#/components/responses/JSONBadRequest"
        "404":
          $ref: "#/components/responses/JSONNotFound"
        "500":
          $ref: "#/components/responses/JSONInternalError"
...
  /api/v1/triggers:
    post:
      summary: Create a saga trigger
      operationId: createTrigger
      tags: [triggers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerCreateRequest"
            example:
              trigger_type: record_transition
              workflow_id: cab_approval_v1
              version: 1
              config:
                record_type: change
                from_state: scheduled
                to_state: pending_cab
                input_mapping: { change_id: "$.record_id" }
              enabled: true
              created_by: admin
      responses:
        "201":
          description: Trigger created. Body is the persisted trigger (note PascalCase field names).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SagaTrigger" }
        "400":
          $ref: "#/components/responses/JSONBadRequest"
        "422":
          description: Type-specific config validation failed. `error` = "invalid_config".
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          $ref: "#/components/responses/JSONInternalError"
    get:
      summary: List saga triggers
      operationId: listTriggers
      tags: [triggers]
      parameters:
        - name: type
          in: query
          description: Filter by trigger type.
          schema: { type: string }
        - name: enabled
          in: query
          description: Filter by enabled flag. Accepts true/false/1/0.
          schema: { type: string, enum: ["true", "false", "1", "0"] }
      responses:
        "200":
          description: The matching triggers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerListResponse"
        "500":
          $ref: "#/components/responses/JSONInternalError"
  /api/v1/triggers/{id}:
    get:
      summary: Get a saga trigger
      operationId: getTrigger
      tags: [triggers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The trigger.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SagaTrigger" }
        "400":
          $ref: "#/components/responses/JSONBadRequest"
        "404":
          description: Trigger not found. `error` = "trigger_not_found".
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          $ref: "#/components/responses/JSONInternalError"
    delete:
      summary: Delete a saga trigger
      operationId: deleteTrigger
      tags: [triggers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Trigger deleted. No body.
        "400":
          $ref: "#/components/responses/JSONBadRequest"
        "404":
          description: Trigger not found. `error` = "trigger_not_found".
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          $ref: "#/components/responses/JSONInternalError"
...
components:
...
  responses:
    JSONBadRequest:
      description: Bad request. `error` = "bad_request".
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    JSONInternalError:
      description: Internal error. `error` = "internal".
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    JSONNotFound:
      description: Not found. `error` = "not_found".
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      description: Structured JSON error envelope used by saga/trigger/workflow endpoints.
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code (e.g. bad_request, saga_not_found, internal).
        message:
          type: string
          description: Human-readable detail.
      required: [error, message]
...
    UserTaskSubmitRequest:
      type: object
      properties:
        submitted_by: { type: string }
        result:
          type: object
          additionalProperties: true
      required: [submitted_by]
...
    TriggerCreateRequest:
      type: object
      properties:
        trigger_type: { type: string }
        workflow_id: { type: string }
        version: { type: integer, minimum: 1 }
        config:
          type: object
          additionalProperties: true
          description: >-
            Shape depends on trigger_type. For "record_transition" the keys
            record_type, from_state, to_state are required non-empty strings.
        enabled: { type: boolean }
        tenant_id:
          type: string
          nullable: true
        created_by: { type: string }
      required: [trigger_type, workflow_id, version, config, created_by]
    SagaTrigger:
      description: >-
        Persisted trigger. The Go domain type carries no JSON tags, so the wire
        representation uses Go field names (PascalCase), as shown here.
      type: object
      properties:
        ID: { type: string, format: uuid }
        TriggerType: { type: string }
        WorkflowID: { type: string }
        Version: { type: integer }
        Config:
          type: object
          additionalProperties: true
        Enabled: { type: boolean }
        TenantID: { type: string, format: uuid, nullable: true }
        CreatedAt: { type: string, format: date-time }
        CreatedBy: { type: string }
      required: [ID, TriggerType, WorkflowID, Version, Config, Enabled, CreatedAt, CreatedBy]
    TriggerListResponse:
      type: object
      properties:
        triggers:
          type: array
          items: { $ref: "#/components/schemas/SagaTrigger" }
      required: [triggers]
...